+++ pac4j-core/src/main/java/org/pac4j/core/util/Pac4jConstants.java
@@ -73,8 +73,7 @@ public interface Pac4jConstants {
     String LOGOUT_URL_PATTERN = "logoutUrlPattern";
 
     /* The default value for the logout url pattern, meaning only relative urls are allowed */
-    /** Constant <code>DEFAULT_LOGOUT_URL_PATTERN_VALUE="^(\\/|\\/[^\\/].*)$"</code> */
-    String DEFAULT_LOGOUT_URL_PATTERN_VALUE = "^(\\/|\\/[^\\/].*)$";
+    String DEFAULT_LOGOUT_URL_PATTERN_VALUE = "^(\\/|\\/[^\\/\\\\].*)$";
 
     /* The config factory parameter */
     /** Constant <code>CONFIG_FACTORY="configFactory"</code> */