+++ documentation/docs/release-notes.md
@@ -10,6 +10,7 @@ title: Release notes&#58;
 - Rework the `SpringResourceLoader` behavior to improve the OIDC/SAML metadata loading resilience:
   - retry every 2 seconds before anything has been loaded
   - retry every 60 seconds after the first loading and return the old data in case of error
+- Fix OpenID federation forced signing of the auth request object for explicitly registered clients when it is not mandatory.
 
 **v6.5.6**:
 - Security fixes/hardenings: